Move the test from test_ask to test_query
